fix genres

+++ db/migrate/20241007060636_drop_genres_table.rb
@@ -1,5 +1,12 @@
-class DropGenresTable < ActiveRecord::Migration[7.1]
-  def change
-    drop_table :genres
+class DropGenresTable < ActiveRecord::Migration[7.0]
+  def up
+    # テーブルが存在する場合のみ削除
+    if table_exists?(:genres)
+      drop_table :genres
+    end
+  end
+
+  def down
+    # ロールバック時の処理（必要な場合）
   end
 end
